Move TaborProgram test
